re: Remove `unwrap_option_or_return` macro (#5938)
There is a macro, which we are using exactly ones. It's better to inline it.
