+++ chain/jsonrpc/src/lib.rs
@@ -243,7 +243,7 @@ impl JsonRpcHandler {
                     >(params)?;
                     self.peer_manager_addr
                         .send(near_network::types::PeerManagerMessageRequest::SetAdvOptions(
-                            near_network::types::SetAdvOptions {
+                            near_network::test_utils::SetAdvOptions {
                                 disable_edge_signature_verification: params
                                     .disable_edge_signature_verification,
                                 disable_edge_propagation: params.disable_edge_propagation,
@@ -263,7 +263,7 @@ impl JsonRpcHandler {
                         near_jsonrpc_adversarial_primitives::SetRoutingTableRequest::parse(params)?;
                     self.peer_manager_addr
                         .send(near_network::types::PeerManagerMessageRequest::SetRoutingTable(
-                            near_network::types::SetRoutingTable {
+                            near_network::test_utils::SetRoutingTable {
                                 add_edges: request.add_edges,
                                 remove_edges: request.remove_edges,
                                 prune_edges: request.prune_edges,
