+++ chain/network/src/peer_manager/peer_manager_actor.rs
@@ -2003,7 +2003,7 @@ impl PeerManagerActor {
     #[perf]
     fn handle_msg_set_adv_options(
         &mut self,
-        msg: crate::types::SetAdvOptions,
+        msg: crate::test_utils::SetAdvOptions,
         _ctx: &mut Context<Self>,
     ) {
         if let Some(disable_edge_propagation) = msg.disable_edge_propagation {
@@ -2025,7 +2025,7 @@ impl PeerManagerActor {
     #[perf]
     fn handle_msg_set_routing_table(
         &mut self,
-        msg: crate::types::SetRoutingTable,
+        msg: crate::test_utils::SetRoutingTable,
         ctx: &mut Context<Self>,
     ) {
         if let Some(add_edges) = msg.add_edges {
