+++ chain/network/src/types.rs
@@ -10,9 +10,8 @@ use crate::private_actix::{
 };
 use crate::routing::routing_table_view::RoutingTableInfo;
 use actix::dev::{MessageResponse, ResponseChannel};
-use actix::{Actor, MailboxError, Message, Recipient};
+use actix::{Actor, MailboxError, Message};
 use futures::future::BoxFuture;
-use futures::FutureExt;
 use near_network_primitives::types::{
     AccountIdOrPeerTrackingShard, AccountOrPeerIdOrHash, Ban, InboundTcpConnect, KnownProducer,
     OutboundTcpConnect, PartialEncodedChunkForwardMsg, PartialEncodedChunkRequestMsg,
@@ -32,23 +31,7 @@ use near_primitives::types::{AccountId, BlockReference, EpochId, ShardId};
 use near_primitives::views::QueryRequest;
 use std::collections::HashMap;
 use std::fmt::Debug;
-use std::sync::RwLock;
 use strum::AsStaticStr;
-
-#[cfg_attr(feature = "deepsize_feature", derive(deepsize::DeepSizeOf))]
-#[derive(Clone, Debug)]
-#[cfg(feature = "test_features")]
-pub struct SetAdvOptions {
-    pub disable_edge_signature_verification: Option<bool>,
-    pub disable_edge_propagation: Option<bool>,
-    pub disable_edge_pruning: Option<bool>,
-    pub set_max_peers: Option<u64>,
-}
-
-#[cfg(feature = "test_features")]
-impl Message for SetAdvOptions {
-    type Result = ();
-}
 
 /// Message from peer to peer manager
 #[derive(strum::AsRefStr, Clone, Debug)]
@@ -113,10 +96,10 @@ pub enum PeerManagerMessageRequest {
     #[cfg(feature = "protocol_feature_routing_exchange_algorithm")]
     StartRoutingTableSync(crate::private_actix::StartRoutingTableSync),
     #[cfg(feature = "test_features")]
-    SetAdvOptions(SetAdvOptions),
+    SetAdvOptions(crate::test_utils::SetAdvOptions),
     #[cfg(feature = "test_features")]
     #[cfg(feature = "protocol_feature_routing_exchange_algorithm")]
-    SetRoutingTable(SetRoutingTable),
+    SetRoutingTable(crate::test_utils::SetRoutingTable),
 }
 
 impl PeerManagerMessageRequest {
@@ -500,53 +483,6 @@ pub trait PeerManagerAdapter: Sync + Send {
     ) -> BoxFuture<'static, Result<PeerManagerMessageResponse, MailboxError>>;
 
     fn do_send(&self, msg: PeerManagerMessageRequest);
-}
-
-#[derive(Default)]
-pub struct NetworkRecipient {
-    peer_manager_recipient: RwLock<Option<Recipient<PeerManagerMessageRequest>>>,
-}
-
-unsafe impl Sync for NetworkRecipient {}
-
-impl NetworkRecipient {
-    pub fn set_recipient(&self, peer_manager_recipient: Recipient<PeerManagerMessageRequest>) {
-        *self.peer_manager_recipient.write().unwrap() = Some(peer_manager_recipient);
-    }
-}
-
-impl PeerManagerAdapter for NetworkRecipient {
-    fn send(
-        &self,
-        msg: PeerManagerMessageRequest,
-    ) -> BoxFuture<'static, Result<PeerManagerMessageResponse, MailboxError>> {
-        self.peer_manager_recipient
-            .read()
-            .unwrap()
-            .as_ref()
-            .expect("Recipient must be set")
-            .send(msg)
-            .boxed()
-    }
-
-    fn do_send(&self, msg: PeerManagerMessageRequest) {
-        let _ = self
-            .peer_manager_recipient
-            .read()
-            .unwrap()
-            .as_ref()
-            .expect("Recipient must be set")
-            .do_send(msg);
-    }
-}
-
-#[cfg_attr(feature = "deepsize_feature", derive(deepsize::DeepSizeOf))]
-#[derive(Message, Clone, Debug)]
-#[rtype(result = "()")]
-pub struct SetRoutingTable {
-    pub add_edges: Option<Vec<Edge>>,
-    pub remove_edges: Option<Vec<SimpleEdge>>,
-    pub prune_edges: Option<bool>,
 }
 
 #[cfg(test)]
