+++ nearcore/src/lib.rs
@@ -15,7 +15,7 @@ use near_client::AdversarialControls;
 use near_client::{start_client, start_view_client, ClientActor, ViewClientActor};
 
 use near_network::routing::start_routing_table_actor;
-use near_network::types::NetworkRecipient;
+use near_network::test_utils::NetworkRecipient;
 use near_network::PeerManagerActor;
 use near_primitives::network::PeerId;
 #[cfg(feature = "rosetta_rpc")]
